+++ src/App.js
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-import PropTypes from 'prop-types';
-import { sortBy } from 'lodash';
-import classNames from 'classnames';
-import './App.css';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faSpinner, faArrowCircleUp, faArrowCircleDown } from '@fortawesome/free-solid-svg-icons'
 
+import './App.css';
+
+import { ButtonWithLoading } from './components/Buttons'
+import Search from './components/Search'
+import Table from './components/Table'
 
 const DEFAULT_QUERY = 'redux';
 const DEFAULT_HPP = '25';
@@ -16,24 +15,6 @@ const PATH_SEARCH = '/search';
 const PARAM_SEARCH = 'query=';
 const PARAM_PAGE = 'page=';
 const PARAM_HPP = 'hitsPerPage=';
-
-const largeColumn = {
-  width: '40%',
-};
-const midColumn = {
-  width: '30%',
-};
-const smallColumn = {
-  width: '10%',
-};
-
-const SORTS = {
-  NONE: list => list,
-  TITLE: list => sortBy(list, 'title'),
-  AUTHOR: list => sortBy(list, 'author'),
-  COMMENTS: list => sortBy(list, 'num_comments').reverse(),
-  POINTS: list => sortBy(list, 'points').reverse(),
-};
 
 const updateSearchTopStoriesState = (hits, page) => (prevState) => {
   const { searchKey, results } = prevState;
@@ -180,235 +161,8 @@ onSearchSubmit(event) {
   }
 }
 
-class Search extends Component {
-  componentDidMount() {
-    if(this.input) {
-      this.input.focus();
-    }
-  }
-
-  render() {
-    const {
-      value,
-      onChange,
-      onSubmit,
-      children
-    } = this.props;
-
-    return (
-    <form onSubmit={onSubmit} >
-      <input
-        type="text"
-        value={value}
-        onChange={onChange}
-        ref={(node) => { this.input = node; }}
-        />
-        <button
-          type="submit">
-          {children}
-        </button>
-    </form>
-    );
-  }
-}
-
-Search.propTypes = {
-  value: PropTypes.string,
-  onChange: PropTypes.func,
-  onSubmit: PropTypes.func,
-  children: PropTypes.node,
-};
-
-
-class Table extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      sortKey: 'NONE',
-      isSortReverse: false,
-    };
-
-    this.onSort = this.onSort.bind(this);
-
-  }
-
-onSort(sortKey) {
-  const isSortReverse = this.state.sortKey === sortKey && !this.state.isSortReverse;
-  this.setState({ sortKey, isSortReverse });
-}
-
-  render() {
-    const {
-      list,
-      onDismiss
-    } = this.props;
-
-    const {
-      sortKey,
-      isSortReverse
-    } = this.state;
-
-    const sortedList = SORTS[sortKey](list);
-    const reverseSortedList = isSortReverse
-      ? sortedList.reverse()
-      : sortedList;
-
-  return (
-
-  <div className="table">
-    <div className="table-header">
-      <span style={{ width: '40%' }}>
-        <Sort
-          sortKey={'TITLE'}
-          onSort={this.onSort}
-          activeSortKey={sortKey}
-          isSortReverse={isSortReverse}
-        >
-          Title
-        </Sort>
-      </span>
-      <span style={{ width: '30%' }}>
-        <Sort
-          sortKey={'AUTHOR'}
-          onSort={this.onSort}
-          activeSortKey={sortKey}
-          isSortReverse={isSortReverse}
-        >
-          Author
-        </Sort>
-      </span>
-      <span style={{ width: '10%' }}>
-        <Sort
-          sortKey={'COMMENTS'}
-          onSort={this.onSort}
-          activeSortKey={sortKey}
-          isSortReverse={isSortReverse}
-        >
-          Comments
-        </Sort>
-      </span>
-      <span style={{ width: '10%' }}>
-        <Sort
-          sortKey={'POINTS'}
-          onSort={this.onSort}
-          activeSortKey={sortKey}
-          isSortReverse={isSortReverse}
-        >
-          Points
-        </Sort>
-      </span>
-      <span style={{ width: '10%' }}>
-        Archive
-      </span>
-    </div>
-    {reverseSortedList.map(item =>
-    <div key={item.objectID} className="table-row">
-    <span style={largeColumn}>
-      <a href={item.url}>{item.title}</a>
-    </span>
-    <span style={midColumn}>{item.author}</span>
-    <span style={smallColumn}>{item.num_comments}</span>
-    <span style={smallColumn}>{item.points}</span>
-    <span style={smallColumn}>
-      <Button 
-      onClick={() => onDismiss(item.objectID)}
-      className="button-inline"
-      >
-      Dismiss
-      </Button>
-    </span>
-    </div>
-    )}
-  </div>
-    );
-  }
-}
-
-Table.propTypes = {
-  list: PropTypes.array.isRequired,
-  onDismiss: PropTypes.func,
-};
-
-const Button = ({
-  onClick, 
-  className, 
-  children
-}) =>
-  <button
-    onClick={onClick}
-    className={className}
-    type="button"
-  >
-    {children}  
-  </button>
-
-Button.propTypes = {
-  onClick: PropTypes.func,
-  className: PropTypes.string,
-  children: PropTypes.node.isRequired,
-};
-
-Button.defaultProps = {
-  className: '',
-};
-
-const Loading = () =>
-<FontAwesomeIcon icon={faSpinner} className="fa-spin" />
-
-const withLoading = (Component) => ({ isLoading, ...rest }) =>
-isLoading
-  ? <Loading />
-  : <Component { ...rest } />
-
-const ButtonWithLoading = withLoading(Button);
-
-const SortActive = ({
-  isSortReverse
-}) =>
-<div>
-{ isSortReverse
-? <FontAwesomeIcon icon={faArrowCircleUp} />
-: <FontAwesomeIcon icon={faArrowCircleDown} />
-}
-</div>
-
-const Sort = ({
-  sortKey,
-  activeSortKey,
-  onSort,
-  isSortReverse,
-  children
-}) => {
-  const sortClass = classNames(
-    'button-inline',
-    { 'button-active': sortKey === activeSortKey }
-  );
-
-  return(
-    <div className = "sorting">
-    <Button 
-      onClick = {() => onSort(sortKey)}
-      className = {sortClass}
-    >
-      {children}
-    </Button>
-    <div>&nbsp;</div>
-    { sortKey === activeSortKey
-    ? <SortActive
-    isSortReverse = {isSortReverse}
-    />
-    : null
-    }
-    </div>
-  );
-}
-
 export default App;
 
 export {
-  Button,
-  Search,
-  Table,
   updateSearchTopStoriesState,
 };
